Add EditorConfig and a very short ESLint
.editorconfig is simple enough that it should explains the rules well.
ESLint, however, has A LOT of configuration options, and I onlu
configure options that I think are important to the coding standards of
the team. I'm open to changes in this ESLint config, especially
single/double quote usage and what counts as an error and what counts as
a warning in ESLint.